requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
Pillow==8.4.0
scipy==1.7.3
six==1.16.0
Werkzeug==2.0.2
